+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/pisdf/checker/HeuristicLoopBreakingDelays.java
@@ -36,6 +36,8 @@
 package org.preesm.algorithm.pisdf.checker;
 
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.HashSet;
 import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
@@ -43,14 +45,16 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.logging.Level;
+import java.util.stream.Collectors;
 import org.jgrapht.alg.cycle.JohnsonSimpleCycles;
 import org.jgrapht.graph.DefaultDirectedGraph;
+import org.preesm.algorithm.pisdf.checker.AbstractGraph.FifoAbstraction;
 import org.preesm.commons.exceptions.PreesmRuntimeException;
 import org.preesm.commons.logger.PreesmLogger;
+import org.preesm.commons.math.MathFunctionsHelper;
 import org.preesm.model.pisdf.AbstractActor;
-import org.preesm.model.pisdf.Delay;
+import org.preesm.model.pisdf.AbstractVertex;
 import org.preesm.model.pisdf.ExecutableActor;
-import org.preesm.model.pisdf.Fifo;
 import org.preesm.model.pisdf.PiGraph;
 import org.preesm.model.pisdf.util.FifoBreakingCycleDetector;
 
@@ -61,11 +65,15 @@ import org.preesm.model.pisdf.util.FifoBreakingCycleDetector;
  */
 public class HeuristicLoopBreakingDelays {
 
-  protected Map<AbstractActor, Integer> actorsNbVisitsTopoRank  = null;
-  protected Map<AbstractActor, Integer> actorsNbVisitsTopoRankT = null;
+  protected final Map<AbstractActor, Integer> actorsNbVisitsTopoRank;
+  protected final Map<AbstractActor, Integer> actorsNbVisitsTopoRankT;
 
-  protected Set<AbstractActor> additionalSourceActors = null;
-  protected Set<AbstractActor> additionalSinkActors   = null;
+  protected final Set<AbstractActor> additionalSourceActors;
+  protected final Set<AbstractActor> additionalSinkActors;
+
+  protected final Map<AbstractVertex, Long>            minCycleBrv;
+  protected final Set<FifoAbstraction>                 breakingFifosAbs;
+  DefaultDirectedGraph<AbstractActor, FifoAbstraction> absGraph;
 
   protected HeuristicLoopBreakingDelays() {
     this.actorsNbVisitsTopoRank = new LinkedHashMap<>();
@@ -73,40 +81,34 @@ public class HeuristicLoopBreakingDelays {
 
     this.additionalSourceActors = new LinkedHashSet<>();
     this.additionalSinkActors = new LinkedHashSet<>();
+
+    this.minCycleBrv = new HashMap<>();
+    this.breakingFifosAbs = new HashSet<>();
+    this.absGraph = null;
   }
 
-  /**
-   * Fifo abstraction to get used in the analysis of this package.
-   *
-   * @author ahonorat
-   */
-  protected class FifoAbstraction {
-    boolean    fullyDelayed;
-    int        nbNonZeroDelays;
-    List<Long> delays;
+  protected void performAnalysis(final PiGraph graph, final Map<AbstractVertex, Long> brv) {
 
-    private FifoAbstraction() {
-      this.fullyDelayed = false;
-      this.nbNonZeroDelays = 0;
-      this.delays = new ArrayList<>();
-    }
-  }
-
-  protected void performAnalysis(final PiGraph graph) {
+    minCycleBrv.putAll(brv);
     // 1. perform flat PiMM to simple JGraphT structure transition.
-    final DefaultDirectedGraph<AbstractActor, FifoAbstraction> absGraph = createAbsGraph(graph);
+    absGraph = AbstractGraph.createAbsGraph(graph);
     // 2. look for cycles
     final JohnsonSimpleCycles<AbstractActor, FifoAbstraction> cycleFinder = new JohnsonSimpleCycles<>(absGraph);
     final List<List<AbstractActor>> cycles = cycleFinder.findSimpleCycles();
 
     // 3 categorize actors in cycle and retrieve fifo with delays breaking loop
     // it is a set since the same fifo can break several cycles if they have some paths in common
-    final Set<FifoAbstraction> breakingFifos = new LinkedHashSet<>();
     for (final List<AbstractActor> cycle : cycles) {
-      final FifoAbstraction breakingFifo = retrieveBreakingFifo(graph, absGraph, cycle);
-      breakingFifos.add(breakingFifo);
-      final AbstractActor src = absGraph.getEdgeSource(breakingFifo);
-      final AbstractActor tgt = absGraph.getEdgeTarget(breakingFifo);
+      final FifoAbstraction breakingFifoAbs = retrieveBreakingFifo(absGraph, cycle);
+      breakingFifosAbs.add(breakingFifoAbs);
+      final AbstractActor src = absGraph.getEdgeSource(breakingFifoAbs);
+      final AbstractActor tgt = absGraph.getEdgeTarget(breakingFifoAbs);
+      long gcdCycle = MathFunctionsHelper.gcd(cycle.stream().map(a -> brv.get(a)).collect(Collectors.toList()));
+      cycle.forEach(a -> {
+        long localBrv = brv.get(a) / gcdCycle;
+        long cycleBrv = minCycleBrv.get(a);
+        minCycleBrv.put(a, Math.min(localBrv, cycleBrv));
+      });
 
       final StringBuilder sb = new StringBuilder(src.getName() + " => " + tgt.getName() + " breaks cycle: ");
       cycle.stream().forEach(a -> sb.append(" -> " + a.getName()));
@@ -120,11 +122,11 @@ public class HeuristicLoopBreakingDelays {
         this.actorsNbVisitsTopoRankT.put(aa, aa.getDataOutputPorts().size());
       }
     }
-    for (final FifoAbstraction breakingFifo : breakingFifos) {
-      final AbstractActor src = absGraph.getEdgeSource(breakingFifo);
-      final AbstractActor tgt = absGraph.getEdgeTarget(breakingFifo);
-      final int newNbVisitsT = this.actorsNbVisitsTopoRankT.get(src) - breakingFifo.nbNonZeroDelays;
-      final int newNbVisits = this.actorsNbVisitsTopoRank.get(tgt) - breakingFifo.nbNonZeroDelays;
+    for (final FifoAbstraction breakingFifoAbs : breakingFifosAbs) {
+      final AbstractActor src = absGraph.getEdgeSource(breakingFifoAbs);
+      final AbstractActor tgt = absGraph.getEdgeTarget(breakingFifoAbs);
+      final int newNbVisitsT = this.actorsNbVisitsTopoRankT.get(src) - breakingFifoAbs.nbNonZeroDelays;
+      final int newNbVisits = this.actorsNbVisitsTopoRank.get(tgt) - breakingFifoAbs.nbNonZeroDelays;
       if ((newNbVisits < 0) || (newNbVisitsT < 0)) {
         throw new PreesmRuntimeException("A loop breaking fifo gave wrong I/O ports number between <" + src.getName()
             + "> and <" + tgt.getName() + ">, leaving.");
@@ -141,8 +143,8 @@ public class HeuristicLoopBreakingDelays {
 
   }
 
-  protected FifoAbstraction retrieveBreakingFifo(final PiGraph graph,
-      final DefaultDirectedGraph<AbstractActor, FifoAbstraction> absGraph, final List<AbstractActor> cycle) {
+  protected FifoAbstraction retrieveBreakingFifo(final DefaultDirectedGraph<AbstractActor, FifoAbstraction> absGraph,
+      final List<AbstractActor> cycle) {
     if (cycle.isEmpty()) {
       throw new PreesmRuntimeException("Bad argument: empty cycle.");
     }
@@ -209,46 +211,6 @@ public class HeuristicLoopBreakingDelays {
     }
 
     return absGraph.getEdge(cycle.get(index), cycle.get((index + 1) % cycle.size()));
-  }
-
-  protected DefaultDirectedGraph<AbstractActor, FifoAbstraction> createAbsGraph(final PiGraph graph) {
-    final DefaultDirectedGraph<AbstractActor,
-        FifoAbstraction> absGraph = new DefaultDirectedGraph<>(FifoAbstraction.class);
-    for (final AbstractActor a : graph.getActors()) {
-      if (a instanceof ExecutableActor) {
-        absGraph.addVertex(a);
-      }
-    }
-    for (final Fifo f : graph.getFifos()) {
-      final AbstractActor absSrc = f.getSourcePort().getContainingActor();
-      final AbstractActor absTgt = f.getTargetPort().getContainingActor();
-      if ((absSrc instanceof ExecutableActor) && (absTgt instanceof ExecutableActor)) {
-        FifoAbstraction fa = absGraph.getEdge(absSrc, absTgt);
-        if (fa == null) {
-          fa = new FifoAbstraction();
-          final boolean res = absGraph.addEdge(absSrc, absTgt, fa);
-          if (!res) {
-            throw new PreesmRuntimeException("Problem while creating graph copy.");
-          }
-        }
-        final Delay d = f.getDelay();
-        if (d == null) {
-          fa.delays.add(0L);
-        } else {
-          fa.nbNonZeroDelays++;
-          fa.delays.add(d.getSizeExpression().evaluate());
-        }
-        boolean fullyDelayed = true;
-        for (final long l : fa.delays) {
-          if (l == 0) {
-            fullyDelayed = false;
-            break;
-          }
-        }
-        fa.fullyDelayed = fullyDelayed;
-      }
-    }
-    return absGraph;
   }
 
 }